+++ neoforge/src/main/java/com/chefmooon/frightsdelight/common/registry/neoforge/FrightsDelightAdvancementsImpl.java
@@ -1,0 +1,21 @@
+package com.chefmooon.frightsdelight.common.registry.neoforge;
+
+import com.chefmooon.frightsdelight.FrightsDelight;
+import net.minecraft.advancements.CriterionTrigger;
+import net.minecraft.core.registries.Registries;
+import net.neoforged.bus.api.IEventBus;
+import net.neoforged.neoforge.registries.DeferredRegister;
+
+import java.util.function.Supplier;
+
+public class FrightsDelightAdvancementsImpl {
+    public static final DeferredRegister<CriterionTrigger<?>> TRIGGERS = DeferredRegister.create(Registries.TRIGGER_TYPE, FrightsDelight.MOD_ID);
+
+    public static <T extends CriterionTrigger<?>> Supplier<T> registerTrigger(String name, Supplier<T> triggerSupplier) {
+        return TRIGGERS.register(name, triggerSupplier);
+    }
+
+    public static void register(IEventBus eventBus) {
+        TRIGGERS.register(eventBus);
+    }
+}