+++ changelog.md
@@ -11,6 +11,9 @@ Minecraft 1.21.5 Fabric Release!
 
 ### Added
 
+- Advancements
+    - 7 New Advancements related to the new food items and blocks
+    - Added 28 items to Master of Fright Advancement
 - New Food
     - Pie:
         - Rotten Flesh Pie (+Slice)